Moved DBConnection class to utils package.
