+++ src/main/java/org/jenkinsci/plugins/environmentdashboard/DashboardBuilder.java
@@ -18,7 +18,7 @@ import javax.servlet.ServletException;
 
 import net.sf.json.JSONObject;
 
-import org.jenkinsci.plugins.model.DBConnection;
+import org.jenkinsci.plugins.environmentdashboard.utils.DBConnection;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
